backend/lightweight/requirements.txt
fastapi==0.109.0
uvicorn==0.27.1
python-multipart
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
asyncpg==0.29.0
redis==5.0.1
python-docx==1.1.0
sse-starlette==1.8.2
weasyprint==60.1
markdown==3.5.1
